+++ src/index.js
@@ -24,6 +24,6 @@ export * from './components/ProgressBar';
 export * from './components/Spinner';
 export * from './components/StatefulButton';
 export * from './components/Tab';
-export * from './components/Navigation';
+export * from './components/VerticalNavigation';
 export * from './decorators';
 export * from './utils';